move slackmsg and slackevent into models
